Add real-data CLI, calibration and proxy events
Introduce interactive support for real-data experiments and an offline calibration workflow, plus proxy event datasets for real evaluation.

Key changes:
- CLI: add menu actions and handlers for Run Real Experiment and Calibrate Scenario; integrate RealBackend and interactive prompts; restrict E2E to synthetic runs and improve run/status printing (date ranges, detector details, model and artifact summaries).
- Calibration: implement cmd_calibrate_scenario to simulate a synthetic experiment, compute empirical summary targets and run the calibration mapping workflow; surface a concise calibration report to the user.
- Data: add proxy event JSON files for gold, SPY, and WTI under data/proxy_events/ (used by real-eval Route A).
- Project config: update .gitignore to ignore data/* but retain data/proxy_events/; add runs_test and add clippy_out.txt.
- Code hygiene and small fixes: add #[allow(...)] pragmas across many modules to suppress dead-code / unused-import warnings; fix a calibration matrix assignment loop and a sign_change_fraction call; several refactors (if-let chaining, printing tweaks, minor logic improvements).
- Docs: update notes/checklist.md to reflect completed real-data and calibration items and other status updates.

These changes enable running and evaluating real-data experiments from the CLI and add the empirical calibration path and example proxy-event datasets to support real-eval workflows.

+++ src/reporting/plot/delay_distribution.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use plotters::prelude::*;
 use serde::{Deserialize, Serialize};
 use std::path::Path;
@@ -21,10 +22,10 @@ pub fn render_delay_distribution(
 
     // Calculate histogram bins
     let max_delay = *input.delays.iter().max().unwrap_or(&1);
-    let mut bins = vec![0usize; max_delay as usize + 1];
+    let mut bins = vec![0usize; max_delay + 1];
     for &delay in &input.delays {
-        if (delay as usize) < bins.len() {
-            bins[delay as usize] += 1;
+        if delay < bins.len() {
+            bins[delay] += 1;
         }
     }
